% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheese.R
\docType{data}
\name{cheese}
\alias{cheese}
\title{Cheese data}
\format{Data frame with columns
\describe{
\item{name, title}{Name of the cheese variety.}
\item{description}{Long description of the cheese.}
\item{href}{Link to page on cheese.com homepage.}
\item{alternative_spelling}{Alternative name of the cheese.}
\item{aroma}{Cheese aroma.}
\item{cakcium_content}{Content of calcium in cheese.}
\item{colour}{Cheese colour}
\item{country_of_origin}{Cheese's country (countries) of origin}
\item{family}{Cheese family.}
\item{fat_content, fat_content_in_dry_matter_}{Fat content in the cheese.}
\item{flavour}{Cheese flavor.}
\item{made_from}{Kind of milk the cheese is made from.}
\item{producers}{Names of cheese producers.}
\item{region}{Region of the country producing cheese.}
\item{rind}{Cheese rind.}
\item{synonyms}{Altenative descriptions for the cheese.}
\item{texture}{Cheese texture}
\item{type}{Cheese type}
\item{vegetarian}{Whether or not the cheese is vegetarian.}
}}
\source{
Worldnews 2018, Specialty cheese catalogue
 <https:://www.cheese.com>
}
\usage{
cheese
}
\description{
Worldwide cheese catalogue from www.cheese.com as of 2018-07-22
}
\keyword{datasets}
