% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheese.R
\docType{data}
\name{cheese_price}
\alias{cheese_price}
\title{Cheese price: Average monthly CME cheese block and barrel prices from www.cheesereporter.com for 2001-2019}
\format{Data frame with columns
\describe{
\item{category}{Either Cheddar Block (40lb) or Cheddar Barrel (500lb).}
\item{date}{Date for the month (BoM) for which data was collected.}
\item{price}{Price in US Dollars.}
}}
\source{
Cheesereporter.com 2019, CME Cheese Block & Barrel Prices
 <https:://www.cheesereporter.com>
}
\usage{
cheese_price
}
\description{
Cheese price: Average monthly CME cheese block and barrel prices from www.cheesereporter.com for 2001-2019
}
\keyword{datasets}
