% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheese.R
\docType{data}
\name{cheese_price_details}
\alias{cheese_price_details}
\title{Cheese price details: CME cheese block and barrel trading details from www.cheesereporter.com for 2011-2018}
\format{Tibble with columns
\describe{
\item{category}{Either Cheddar Block (40lb) or Cheddar Barrel (500lb).}
\item{year, month, day}{Date when the trading event happened.}
\item{price}{Price in US Dollars.}
\item{volume}{Number of trades happening on that day.}
\item{deals}{List-column storing number of trades, price and bid/offer flag for transaction.}
\item{bids}{List-column storing number of unfilled bids, price and bid/offer flag.}
\item{offers}{List-column storing number of unfilled offers, price and bid/offer flag.}
}}
\source{
Cheesereporter.com 2019, CME Cheese Block & Barrel Trading Details
 <https:://www.cheesereporter.com>
}
\usage{
cheese_price_details
}
\description{
Cheese price details: CME cheese block and barrel trading details from www.cheesereporter.com for 2011-2018
}
\keyword{datasets}
