% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl.R
\name{crawl_cheese_com}
\alias{crawl_cheese_com}
\title{Crawl www.cheese.com and return tidy dataframe with information about world cheeses}
\usage{
crawl_cheese_com(path = "www.cheese.com")
}
\arguments{
\item{path}{address of cheese.com website}
}
\value{

}
\description{
Crawl www.cheese.com and return tidy dataframe with information about world cheeses
}
\examples{
crawl_cheese_com()
}
