\docType{data}
\name{heart_disease}
\alias{heart_disease}
\title{Heart Disease}
\format{See "Source" for link to dataset home page}
\usage{
heart_disease
}
\description{
This is a cleaned up version of the "heart disease data set" found in the UCI Machine Learning Repository (https://archive.ics.uci.edu/ml/datasets/Heart+Disease), containing a subset of the default variables.
}
\source{https://archive.ics.uci.edu/ml/datasets/Heart+Disease}
\keyword{datasets}